+++ src/js/components/Facebook/Main.js
@@ -1,5 +1,5 @@
 import React from "react";
-import FacebookActionCreators from "../../actions/FacebookActionCreators";
+import FacebookActions from "../../actions/FacebookActions";
 import FacebookStore from "../../stores/FacebookStore";
 import FacebookDownloadPicture from "../../components/Facebook/FacebookDownloadPicture";
 import FacebookPicture from "../../components/Facebook/FacebookPicture";
@@ -22,7 +22,7 @@ class Main extends React.Component {
     }
 
     componentDidMount () {
-        FacebookActionCreators.initFacebook();
+        FacebookActions.initFacebook();
         this.listener = FacebookStore.addListener(this._onChange.bind(this));
         this.voterListener = VoterStore.addListener(this._onChange.bind(this));
     }
